requirements.txt
PyMuPDF>=1.24
gradio>=4.0
pydantic>=2.0
tenacity>=8.0
python-dotenv>=1.0
